requirements.txt
apturl==0.5.2
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
autopep8==2.0.1
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2022.9.14
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
Django==4.1.2
django-filter==23.1
djangorestframework==3.14.0
duplicity==0.8.21
fasteners==0.14.1
filelock==3.6.0
Flask==2.2.3
future==0.18.2
h11==0.13.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.3
MarkupSafe==2.1.2
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
outcome==1.2.0
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.1
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycodestyle==2.10.0
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
systemd-python==234
tomli==2.0.1
trio==0.21.0
trio-websocket==0.9.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.13.0+ds
wadllib==1.3.6
Werkzeug==2.2.3
wsproto==1.2.0
xdg==5
xkit==0.0.0
zipp==1.0.0
